Refactor the code to extract out closely related code as classes.
