+++ src/main/java/DukeFileNotFoundException.java
@@ -1,0 +1,5 @@
+public class DukeFileNotFoundException extends DukeException {
+    public DukeFileNotFoundException() {
+        super("No available file");
+    }
+}
